+++ owoify/utility/presets.py
@@ -3,8 +3,11 @@ from owoify.utility.mapping import *
 
 SPECIFIC_WORD_MAPPING_LIST = [
     map_fuc_to_fwuc, map_mom_to_mwom, map_time_to_tim, map_me_to_mwe,
-    map_n_vowel_to_ny, map_over_to_owor, map_ove_to_uv, map_haha_to_hehe_xd,
-    map_the_to_teh, map_you_to_u, map_read_to_wead, map_worse_to_wose
+    map_over_to_owor, map_ove_to_uv, map_haha_to_hehe_xd,
+    map_the_to_teh, map_you_to_u, map_read_to_wead, map_worse_to_wose,
+    map_great_to_gwate, map_aviat_to_awiat, map_dedicat_to_deditat,
+    map_remember_to_rember, map_when_to_wen, map_frightened_to_frigten,
+    map_meme_to_mem, map_feel_to_fell
 ]
 
 UVU_MAPPING_LIST = [
@@ -19,9 +22,9 @@ UWU_MAPPING_LIST = [
 ]
 
 OWO_MAPPING_LIST = [
-    map_ll_to_ww, map_vowel_or_r_except_o_l_to_wl, map_old_to_owld,
+    map_n_vowel_to_ny, map_ll_to_ww, map_vowel_or_r_except_o_l_to_wl, map_old_to_owld,
     map_ol_to_owl, map_l_or_r_o_to_wo, map_specific_consonants_o_to_letter_and_wo,
     map_v_or_w_le_to_wal, map_fi_to_fwi, map_ver_to_wer, map_poi_to_pwoi,
     map_specific_consonants_le_to_letter_and_wal, map_consonant_r_to_consonant_w,
-    map_ly_to_wy, map_ple_to_pwe, map_nr_to_nw
+    map_ly_to_wy, map_ple_to_pwe, map_nr_to_nw, map_mem_to_mwem, unmap_nywo_to_nyo
 ]